+++ src/main/clojure/com/github/clojure_lsp/intellij/action/refactors.clj
@@ -1,9 +1,9 @@
 (ns com.github.clojure-lsp.intellij.action.refactors
   (:require
-   [com.github.clojure-lsp.intellij.client :as lsp-client]
-   [com.github.clojure-lsp.intellij.db :as db]
    [camel-snake-kebab.core :as csk]
-   [com.github.clojure-lsp.intellij.editor :as editor])
+   [com.github.clojure-lsp.intellij.client :as lsp-client]
+   [com.github.clojure-lsp.intellij.editor :as editor]
+   [com.github.clojure-lsp.intellij.server :as server])
   (:import
    [com.intellij.openapi.actionSystem AnActionEvent]
    [com.intellij.openapi.actionSystem CommonDataKeys]
@@ -12,8 +12,7 @@
 (set! *warn-on-reflection* true)
 
 (defn ^:private action-performed [command _ ^AnActionEvent event]
-  (when-let [client (and (identical? :connected (:status @db/db*))
-                         (:client @db/db*))]
+  (when-let [client (server/connected-client)]
     (when-let [editor ^Editor (.getData event CommonDataKeys/EDITOR)]
       (let [[line character] (editor/editor->cursor-position editor)]
         (lsp-client/request! client [:workspace/executeCommand