+++ src/main/clojure/com/github/clojure_lsp/intellij/extension/code_action.clj
@@ -4,7 +4,8 @@
    [clojure.core.memoize :as memoize]
    [com.github.clojure-lsp.intellij.client :as lsp-client]
    [com.github.clojure-lsp.intellij.db :as db]
-   [com.github.clojure-lsp.intellij.editor :as editor])
+   [com.github.clojure-lsp.intellij.editor :as editor]
+   [com.github.clojure-lsp.intellij.server :as server])
   (:import
    [com.intellij.openapi.editor Editor]
    [com.intellij.openapi.project Project]
@@ -15,8 +16,7 @@
 (defonce code-action-name->title* (atom {}))
 
 (defn ^:private req-code-actions [uri [line character]]
-  (when-let [client (and (identical? :connected (:status @db/db*))
-                         (:client @db/db*))]
+  (when-let [client (server/connected-client)]
     (let [diagnostics (->> (get-in @db/db* [:diagnostics uri])
                            (filterv (fn [{{:keys [start end]} :range}]
                                       (and (<= (:line start) line (:line end))